// src/data-interfaces/post.user.interface.ts
import { IGeo } from './put.user.interface';

export interface IcreateUser {
  name: string;
  username: string;
  email: string;
  address?: {
    street?: string;
    suite?: string;
    city?: string;
    zipcode?: string;
    geo?: IGeo;
  };
  phone: string;
  website?: string;
  company?: {
    name?: string;
    catchPhrase?: string;
    bs?: string;
  };
}